+++ backend/tests/conftest.py
@@ -1,10 +1,7 @@
 from __future__ import print_function
 
-import sys
-
 import psycopg2
 import pytest
-from django.db import connections
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
 
@@ -63,52 +60,3 @@ def run_non_transactional_raw_sql(sqls, dbinfo):
         cursor.execute(sql)
 
     conn.close()
-
-
-# Nicest way of printing to stderr sourced from
-# https://stackoverflow.com/questions/5574702/how-to-print-to-stderr-in-python
-def eprint(*args, **kwargs):
-    print(*args, file=sys.stderr, **kwargs)
-
-
-@pytest.fixture()
-def user_tables_in_separate_db(settings):
-    """
-    Creates a temporary database and sets up baserow so it is used to store user tables.
-
-    Currently this has only been implemented at a function level scope as adding
-    databases to settings.DATABASES causes pytest to assume they are extra replica dbs
-    and spend ages setting them up as mirrors. Instead keeping this at the functional
-    scope lets us keep it simple and quick.
-    """
-
-    default_db = settings.DATABASES["default"]
-    user_table_db_name = f'{default_db["NAME"]}_user_tables'
-
-    # Print to stderr to match pytest-django's behaviour for logging about test
-    # database setup and teardown.
-    eprint(f"Dropping and recreating {user_table_db_name} for test.")
-
-    settings.USER_TABLE_DATABASE = "user_tables_database"
-    settings.DATABASES["user_tables_database"] = dict(default_db)
-    settings.DATABASES["user_tables_database"]["NAME"] = user_table_db_name
-
-    # You cannot drop databases inside transactions and django provides no easy way
-    # of turning them off temporarily. Instead we need to open our own connection so
-    # we can turn off transactions to perform the required setup/teardown sql. See:
-    # https://pytest-django.readthedocs.io/en/latest/database.html#using-a-template
-    # -database-for-tests
-    run_non_transactional_raw_sql(
-        [
-            f"DROP DATABASE IF EXISTS {user_table_db_name}; ",
-            f"CREATE DATABASE {user_table_db_name}",
-        ],
-        default_db,
-    )
-
-    yield connections["user_tables_database"]
-
-    # Close django's connection to the user table db so we can drop it.
-    connections["user_tables_database"].close()
-
-    run_non_transactional_raw_sql([f"DROP DATABASE {user_table_db_name}"], default_db)