+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* The internal setting allowing Baserow to run with the user tables in a separate 
+  database has been removed entirely to prevent data integrity issues.
+  
 ## Released (2021-07-16)
 
 * Fix bug preventing fields not being able to be converted to link row fields in some